requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
click==8.1.3
docutils==0.18.1
Flask==2.1.2
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
graphviz==0.13.2
htmldocx==0.0.6
Jinja2==3.1.2
jsonpath-ng==1.5.3
jsonschema==4.6.1
MarkupSafe==2.1.1
marshmallow==3.17.0
requests==2.28.1
robotframework==5.0.1
robotframework-pythonlibcore==3.0.0
rpaframework==15.1.4
rpaframework-core==9.0.1
Werkzeug==0.16.1
gunicorn==20.1.0
